User Credentials updated

+++ src/app/views/pages/auth/principal-login/principal-login.component.ts
@@ -13,10 +13,14 @@ export class PrincipalLoginComponent implements OnInit {
 
   returnUrl: any;
   persalnumber: any;
+  password: any;
+  confirmpassword: any;
   idnumber: any;
   myStorage: any;
   response: any;
   error;
+  passworderror;
+  confirmerror;
 
 
 
@@ -26,7 +30,9 @@ export class PrincipalLoginComponent implements OnInit {
 
   ngOnInit(): void {
     this.form = this.formBuilder.group({
-      persalnumber: ""
+      persalnumber: "",
+      password: "",
+      confirmpassword: ""
     });
  this.error = false;
   }
@@ -34,7 +40,11 @@ export class PrincipalLoginComponent implements OnInit {
   getOtp() {
 
     this.persalnumber = this.form.controls['persalnumber'].value;
+    this.password = this.form.controls['password'].value;
+    this.confirmpassword = this.form.controls['confirmpassword'].value;
     this.error = false;
+    this.passworderror = false;
+    this.confirmerror = false;
 
 //     if(this.persalnumber  != "") {
 
@@ -61,21 +71,37 @@ export class PrincipalLoginComponent implements OnInit {
 
     // })
 
-    this._principal.sendUserOTP(this.persalnumber).subscribe((res: any) => {
-      this.response = res;
-      console.log(res.length);
+    if (this.password.length > 8) {
+      
+      if (this.password == this.confirmpassword) {
+        
+        this._principal.sendUserOTP(this.persalnumber).subscribe((res: any) => {
+          this.response = res;
+          console.log(res.length);
+    
+          if(res.length > 1){        
+          localStorage.setItem('cellnumber', res);
+          localStorage.setItem('Idnumber', this.persalnumber);
+          localStorage.setItem('Pass', this.password);
+          this.myStorage = localStorage.getItem("cellnumber");
+          this.router.navigate(['/auth/verify-otp'])
+          } else {
+            console.log("error")
+            this.error = true;      }
+    
+        })
 
-      if(res.length > 1){        
-      localStorage.setItem('cellnumber', res);
-      // localStorage.setItem('Persal', this.persalnumber);
-      localStorage.setItem('Idnumber', this.persalnumber);
-      this.myStorage = localStorage.getItem("cellnumber");
-      this.router.navigate(['/auth/verify-otp'])
       } else {
-        console.log("error")
-        this.error = true;      }
+        console.log("confirmerror")
+        this.confirmerror = true;
+      }
 
-    })
+    } else {
+      console.log("passworderror")
+         this.passworderror = true;
+    }
+
+ 
 
   }
